backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
qdrant-client==1.7.1
sentence-transformers==2.2.2
python-multipart==0.0.6
markdown==3.5.1
google-generativeai==0.3.2
